3.2b4.zip
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@4319 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ printView.m
@@ -130,7 +130,7 @@
 
 //---------------------------------------------------------------------------
  
-- (NSRect)rectForPage:(int)page
+- (NSRect)rectForPage:(NSInteger)page
 {
 	//Before printing each page, the pagination machinery sends the view a rectForPage: message. 
 	//Your implementation of rectForPage: should use the supplied page number and the current printing information to
